requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pandas>=2.2.0
tushare>=1.4.0
ollama>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.7.0
pytest>=8.0.0
matplotlib>=3.9.0
mplfinance>=0.12.10b0
